% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abruzzo.r
\docType{data}
\name{abruzzo}
\alias{abruzzo}
\title{Oggetto sf della regione abruzzo}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
abruzzo
}
\description{
Oggetto sf della regione abruzzo
}
\keyword{datasets}
