% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilicata.r
\docType{data}
\name{basilicata}
\alias{basilicata}
\title{Oggetto sf della regione basilicata}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
basilicata
}
\description{
Oggetto sf della regione basilicata
}
\keyword{datasets}
