% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calabria.r
\docType{data}
\name{calabria}
\alias{calabria}
\title{Oggetto sf della regione calabria}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
calabria
}
\description{
Oggetto sf della regione calabria
}
\keyword{datasets}
