% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campania.r
\docType{data}
\name{campania}
\alias{campania}
\title{Oggetto sf della regione campania}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
campania
}
\description{
Oggetto sf della regione campania
}
\keyword{datasets}
