% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emiliaromagna.r
\docType{data}
\name{emiliaromagna}
\alias{emiliaromagna}
\title{Oggetto sf della regione emiliaromagna}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
emiliaromagna
}
\description{
Oggetto sf della regione emiliaromagna
}
\keyword{datasets}
