% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friuliveneziagiulia.r
\docType{data}
\name{friuliveneziagiulia}
\alias{friuliveneziagiulia}
\title{Oggetto sf della regione friuliveneziagiulia}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
friuliveneziagiulia
}
\description{
Oggetto sf della regione friuliveneziagiulia
}
\keyword{datasets}
