% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italia.r
\docType{data}
\name{italia}
\alias{italia}
\title{Oggetto sf della regione italia}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 101 columns.
}
\usage{
italia
}
\description{
Oggetto sf della regione italia
}
\keyword{datasets}
