% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italia_senza_isole.r
\docType{data}
\name{italia_senza_isole}
\alias{italia_senza_isole}
\title{Oggetto sf della regione italia_senza_isole}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 20 rows and 11 columns.
}
\usage{
italia_senza_isole
}
\description{
Oggetto sf della regione italia_senza_isole
}
\keyword{datasets}
