% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazio.r
\docType{data}
\name{lazio}
\alias{lazio}
\title{Oggetto sf della regione lazio}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
lazio
}
\description{
Oggetto sf della regione lazio
}
\keyword{datasets}
