% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liguria.r
\docType{data}
\name{liguria}
\alias{liguria}
\title{Oggetto sf della regione liguria}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
liguria
}
\description{
Oggetto sf della regione liguria
}
\keyword{datasets}
