% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lombardia.r
\docType{data}
\name{lombardia}
\alias{lombardia}
\title{Oggetto sf della regione lombardia}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
lombardia
}
\description{
Oggetto sf della regione lombardia
}
\keyword{datasets}
