% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marche.r
\docType{data}
\name{marche}
\alias{marche}
\title{Oggetto sf della regione marche}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
marche
}
\description{
Oggetto sf della regione marche
}
\keyword{datasets}
