% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molise.r
\docType{data}
\name{molise}
\alias{molise}
\title{Oggetto sf della regione molise}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
molise
}
\description{
Oggetto sf della regione molise
}
\keyword{datasets}
