% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piemonte.r
\docType{data}
\name{piemonte}
\alias{piemonte}
\title{Oggetto sf della regione piemonte}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
piemonte
}
\description{
Oggetto sf della regione piemonte
}
\keyword{datasets}
