% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puglia.r
\docType{data}
\name{puglia}
\alias{puglia}
\title{Oggetto sf della regione puglia}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
puglia
}
\description{
Oggetto sf della regione puglia
}
\keyword{datasets}
