% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sardegna.r
\docType{data}
\name{sardegna}
\alias{sardegna}
\title{Oggetto sf della regione sardegna}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
sardegna
}
\description{
Oggetto sf della regione sardegna
}
\keyword{datasets}
