% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicilia.r
\docType{data}
\name{sicilia}
\alias{sicilia}
\title{Oggetto sf della regione sicilia}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
sicilia
}
\description{
Oggetto sf della regione sicilia
}
\keyword{datasets}
