% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toscana.r
\docType{data}
\name{toscana}
\alias{toscana}
\title{Oggetto sf della regione toscana}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
toscana
}
\description{
Oggetto sf della regione toscana
}
\keyword{datasets}
