% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trentino.r
\docType{data}
\name{trentino}
\alias{trentino}
\title{Oggetto sf della regione trentino}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
trentino
}
\description{
Oggetto sf della regione trentino
}
\keyword{datasets}
