% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbria.r
\docType{data}
\name{umbria}
\alias{umbria}
\title{Oggetto sf della regione umbria}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
umbria
}
\description{
Oggetto sf della regione umbria
}
\keyword{datasets}
