% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleaosta.r
\docType{data}
\name{valleaosta}
\alias{valleaosta}
\title{Oggetto sf della regione valleaosta}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
valleaosta
}
\description{
Oggetto sf della regione valleaosta
}
\keyword{datasets}
