% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veneto.r
\docType{data}
\name{veneto}
\alias{veneto}
\title{Oggetto sf della regione veneto}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\usage{
veneto
}
\description{
Oggetto sf della regione veneto
}
\keyword{datasets}
